# Lab07/src/main/resources/openapi/internetService.yaml
openapi: 3.0.3
info:
  title: Internet Provider Service API
  version: 1.0.0
  description: |
    REST API do zarządzania klientami, subskrypcjami, cennikiem,
    fakturami i wpłatami dla dostawcy usług internetowych.
servers:
  - url: http://localhost:8080/api
...
components:
  schemas:
...
    ClientDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        clientNumber:
          type: string
      required:
        - firstName
        - lastName
        - clientNumber
...
paths:
  /clients:
    get:
      tags: [Clients]
      summary: Pobierz wszystkich klientów
      responses:
        '200':
          description: Lista klientów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientDTO'
    post:
      tags: [Clients]
      summary: Dodaj nowego klienta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        '201':
          description: Klient utworzony
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
  /clients/{id}:
    get:
      tags: [Clients]
      summary: Pobierz klienta po ID
      parameters:
        - name: id
          in: path
          description: ID klienta
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Klient znaleziony
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
        '404':
          description: Klient nie znaleziony
...